Move walletSlice to web3 feature
